Solved PR comments.
